Fix goto definition
The new OutlineMode introduced for pyQode/pyQode#24 has a list of ``Definition`` instead of a list of ``Name`` and the find node method did not exists on ``Definition`` causing an AttributeError to be raised whenever the user wants to go to the definition.

Also fix a non-critical old bug (not due to the new OutlineMode) when there is not definition to go to.